move cwl converter under management
